// src/typings/db.ts
export interface IProduct extends IProductOrigin {
  joinMembr: string[];
  creator: {
    _id: string;
    name: string;
    email: string;
    phone: number;
  };
}

export interface IProductType2 extends IProductOrigin {
  joinMembr: string[];
  creator: string;
}

export interface ICurriculum {
  _id: string;
  title: string;
  lessons: ILesson[];
}

export interface ILesson {
  _id: string;
  title: string;
  content: string;
  mediaId: string;
  mediaTime?: number;
  filename: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface IProductList {
  products: IProduct[];
  productsCount: number;
}

export interface IinfinityProduct {
  pageParams: number[];
  pages: [
    {
      products: IProduct[];
      nextPage: number;
      isLast: boolean;
    }
  ];
}

export interface INotice {
  _id: string;
  category?: string;
  title: string;
  body?: string;
  imgurl?: string;
  summary: string;
  updatedAt: string;
}
// ...
export interface ISSR {
  SsrData: {
    products: IProductList;
    blogData: INotice[];
    noticeData: INotice[];
  };
}